+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/tabs/ManageTabs/manageTabs.module.scss
@@ -1,0 +1,14 @@
+.manage {
+    button {
+        display: flex;
+        align-items: center;
+        gap: 0.5rem;
+        margin-top: -0.6rem;
+        text-decoration: none;
+    }
+}
+.modalHeader {
+    h3 {
+        margin: 0;
+    }
+}
